apps: mchpcam: remove warnings while initializing controls
Remove the warning message while initializing the default values for v4l2
controls. If the user doesn't pass the value through arguments the default
value of the control is received from the pipeline handler. This warning
message is printed if the app couldn't read the value from the pipeline
handler. However the values of controls that are not passed from the
application will retain its value that is already set by the driver or
application that is invoked before.

Signed-off-by: Balamanikandan Gunasundar <balamanikandan.gunasundar@microchip.com>